+++ src/Component/Concerns/ConfigurableComponent.php
@@ -7,7 +7,7 @@ namespace CraftCms\Cms\Component\Concerns;
 use BackedEnum;
 use CraftCms\Cms\Component\Contracts\ConfigurableComponentInterface;
 use CraftCms\Cms\Component\Events\DefineSettingsAttributes;
-use CraftCms\Cms\Cp\Forms\Form;
+use CraftCms\Cms\Cp\Forms\Contracts\FormDefinition;
 use CraftCms\Cms\Support\DateTimeHelper;
 use CraftCms\Cms\Support\Str;
 use CraftCms\Cms\Support\Utils;
@@ -75,7 +75,7 @@ trait ConfigurableComponent
         return $settings;
     }
 
-    public function getSettingsForm(bool $readOnly): ?Form
+    public function getSettingsForm(bool $readOnly): ?FormDefinition
     {
         return null;
     }